+++ graflo/hq/__init__.py
@@ -5,16 +5,20 @@ multiple components for graph database operations.
 """
 
 from graflo.hq.caster import Caster, IngestionParams
+from graflo.hq.db_writer import DBWriter
 from graflo.hq.graph_engine import GraphEngine
 from graflo.hq.inferencer import InferenceManager
+from graflo.hq.registry_builder import RegistryBuilder
 from graflo.hq.resource_mapper import ResourceMapper
 from graflo.hq.sanitizer import SchemaSanitizer
 
 __all__ = [
     "Caster",
+    "DBWriter",
     "GraphEngine",
     "IngestionParams",
     "InferenceManager",
+    "RegistryBuilder",
     "ResourceMapper",
     "SchemaSanitizer",
 ]