+++ graflo/hq/graph_engine.py
@@ -7,7 +7,7 @@ and data ingestion.
 
 import logging
 
-from graflo import Schema
+from graflo.architecture.schema import Schema
 from graflo.onto import DBType
 from graflo.architecture.onto_sql import SchemaIntrospectionResult
 from graflo.db import ConnectionManager, PostgresConnection